requirements.txt
pytest==7.4.2
pytest-html==3.2.0
requests==2.31.0
kafka-python==2.0.2
redis==5.3.1
pymongo==4.5.0
docker==6.1.3
boto3==1.42.13
google-cloud-storage==2.12.0
selenium==4.39.0
PyYAML==6.0.3
